requirements.txt
flask>=2.2.5
requests>=2.31.0
openai>=1.3.5
python-dotenv>=1.0.0
gunicorn>=21.2.0